Extract/src/stock_market_extract/requirements.txt
yfinance~=0.2.44
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1
numpy~=1.26.4
requests~=2.32.3
oauthlib~=3.2.2
pandas~=2.2.2
transformers~=4.45.2
psycopg2-binary~=2.9.6